[driver/glfw] split glDriver#CreateWindow()
Instead of runtime OS check it is now split by build flags.
The method has also moved into the driver files because it’s a driver
method.
